feat: create inngest client and function
Replace with the current Redis pub sub
