+++ apps/server/src/index.ts
@@ -1,18 +1,17 @@
 import FiisRoutes from "./modules/fiis/routes";
 import FiagroRoutes from "./modules/fiagro/routes";
 import NotionRoutes from "./modules/notion/routes";
-import { onServerStarts } from "./utils/server";
 import { DEFAULT_PORT } from "./constants";
 import { parseNumber } from "./utils/numbers";
+import inngestPlugin from "./plugins/inngest";
 import Elysia from "elysia";
 
 const app = new Elysia()
+app.use(inngestPlugin)
 
 app.use(FiisRoutes)
 app.use(FiagroRoutes)
 app.use(NotionRoutes)
-
-onServerStarts();
 
 const port = parseNumber(process.env.PORT) ?? DEFAULT_PORT;
 app.listen(port)
