+++ apps/server/src/services/inngest/index.ts
@@ -1,0 +1,3 @@
+import { Inngest } from "inngest";
+
+export const inngest = new Inngest({ id: "stock-notion-server" });
