--- apps/server/src/utils/server.ts
@@ -1,5 +1,0 @@
-import { subscribeNotionDataSourceFi } from "@/services/notion/channel";
-
-export function onServerStarts() {
-  subscribeNotionDataSourceFi();
-}
